misc refactor: naming, structure
