+++ graphql/mutations/createClipboard.js
